medical_diagnosis_assistant/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.1
djangorestframework==3.15.2
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
Markdown==2.6.10
MarkupSafe==2.1.5
pillow==10.4.0
Pygments==2.2.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
